% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackjack-package.R, R/deal_hand.R
\name{blackjack-package}
\alias{blackjack-package}
\title{Blackjack Package}
\description{
The `blackjack` package provides tools to simulate the game of Blackjack in R.
It supports single and multiplayer modes, dealer logic, and strategy testing.
}
