% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_class.R
\name{card}
\alias{card}
\title{Create a Card Vector}
\usage{
card(rank, suit)
}
\arguments{
\item{rank}{Character vector of card ranks (e.g., "A", "2", ..., "K").}

\item{suit}{Character vector of suits (e.g., "♠", "♥", "♦", "♣").}
}
\value{
An object of class `card`.
}
\description{
Constructs a custom card vector object using vctrs, representing card ranks and suits.
}
