% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shuffle_deck.R
\name{create_shuffled_deck}
\alias{create_shuffled_deck}
\title{Create a Shuffled Blackjack Shoe}
\usage{
create_shuffled_deck(noOfDecks = 4)
}
\arguments{
\item{noOfDecks}{Number of decks to include in the shoe (default = 4)}
}
\value{
A character vector of 208 shuffled cards
}
\description{
This function creates a shuffled deck of cards for Blackjack, consisting of multiple standard decks combined.
}
