% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealer_turn.R
\name{dealer_turn}
\alias{dealer_turn}
\title{Dealer Turn Logic}
\usage{
dealer_turn(dealer_hand, deck)
}
\arguments{
\item{dealer_hand}{The dealer's current hand}

\item{deck}{The current deck of cards}
}
\value{
A list with:
\describe{
  \item{hand}{The final dealer hand after drawing cards.}
  \item{deck}{The updated deck after removing drawn cards.}
  \item{total}{The final total value of the dealer's hand.}
}
}
\description{
Simulates the dealer's decisions in Blackjack.
}
