% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_turn.R
\name{player_turn}
\alias{player_turn}
\title{Player Turn Interaction}
\usage{
player_turn(player_hand, deck, input_fn = readline)
}
\arguments{
\item{player_hand}{A vector of cards representing the player's hand.}

\item{deck}{A vector of cards representing the remaining deck.}
}
\value{
A list containing:
\item{hand}{The updated player hand after the turn.}
\item{deck}{The updated deck after the player's turn.}
\item{total}{The total value of the player's hand.}
}
\description{
Handles the player's turn in a Blackjack game, allowing them to hit or stand.
}
