% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_class.R
\name{card_is_face}
\alias{card_is_face}
\title{Check if Card is a Face Card}
\usage{
card_is_face(x)
}
\arguments{
\item{x}{A card object.}
}
\value{
A logical vector: True for face cards, False otherwise.#'
}
\description{
Determines if the card is a face card (Jack, Queen or King)
}
