% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_class.R
\name{card_rank}
\alias{card_rank}
\title{Get Rank of Card}
\usage{
card_rank(x)
}
\arguments{
\item{x}{A card object.}
}
\value{
A character vector of card ranks.
}
\description{
Extracts the rank from a card object.
}
