% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_class.R
\name{card_suit}
\alias{card_suit}
\title{Get Suit of Card}
\usage{
card_suit(x)
}
\arguments{
\item{x}{A card object.}
}
\value{
A character vector of card suits.
}
\description{
Extracts the suit from a card object.
}
