% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_value.R
\name{card_value}
\alias{card_value}
\title{Score a Blackjack Hand}
\usage{
card_value(cards)
}
\arguments{
\item{cards}{A character vector like "A♠", "10♥"}
}
\value{
Total hand value (numeric)
}
\description{
Calculates total hand value, adjusting for Aces
}
