% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_blackjack.R
\name{play_blackjack}
\alias{play_blackjack}
\title{Play One Round of Blackjack with Action}
\usage{
play_blackjack(..., input_fn = readline)
}
\arguments{
\item{n_players}{Number of players (default is 1)}
}
\value{
A list containing the player's hand, dealer's hand, scores, and the result of the game.
}
\description{
This function simulates a full round of Blackjack, allowing the player to hit or stand.
}
