% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_blackjack_multi.R
\name{play_blackjack_multi}
\alias{play_blackjack_multi}
\title{Simulate Multiplayer Blackjack (2 players)}
\usage{
play_blackjack_multi(n_players = 2, ..., input_fn = readline)
}
\arguments{
\item{n_players}{Number of players (default is 2)}
}
\value{
A data frame containing the results of the game, including each player's hand, score, and result against the dealer
}
\description{
This function simulates a round of Blackjack for two players against a dealer.
}
