% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_blackjack_multi.R
\name{play_blackjack_multi}
\alias{play_blackjack_multi}
\title{Simulate Multiplayer Blackjack (2 players)}
\usage{
play_blackjack_multi(n_players = 2, ..., input_fn = readline)
}
\arguments{
\item{n_players}{Number of players (default is 2)}

\item{...}{Additional arguments (currently unused; reserved for future extensions)}

\item{input_fn}{holder to obtain player actions (defaults to \code{readline}) which can be overridden to inject fake input for unit tests or if the user want to add fake actions.}
}
\value{
A data frame containing the results of the game, including each player's hand, score, and result against the dealer
}
\description{
This function simulates a round of Blackjack for two players against a dealer.
}
